Add git-sparse command.
I don't like storing this in bin, but it's probably the least bad place.

Other options include .local/stow, but then (a) I'd have to restow on any major
changes (possibly able to be done automatically with a
post-checkout/post-rewrite hook) and (b) .local/stow may not make sense on all
platforms. Finally, (c) if these get big, it should be it's own repo (possibly
a use of git submodules, despite their awkwardness).